cat-activities-analysis-api/app.py
"""
Example external analysis service for cat activity detection.

This is a sample implementation showing how you could build an external service
that the Home Assistant integration can call to analyze images.

You can run this as a separate FastAPI service or integrate it with cloud AI services.
"""

from fastapi import FastAPI, File, UploadFile, HTTPException
from fastapi.responses import JSONResponse
import cv2
import numpy as np
from PIL import Image
import io
from datetime import datetime
from typing import Dict, Any
import logging

# Optional: Import AI/ML libraries for more sophisticated analysis
# import torch  # For PyTorch models
# import tensorflow as tf  # For TensorFlow models
# from ultralytics import YOLO  # For YOLO object detection

app = FastAPI(title="Cat Activities Analysis Service", version="1.0.0")

# Set up logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Load your AI model here (example)
# model = YOLO('yolov8n.pt')  # Pre-trained YOLO model
# or load your custom trained model for cat detection

CAT_ACTIVITIES = [
    "sleeping", "eating", "playing", "grooming", 
    "exploring", "sitting", "lying_down", "walking", "unknown"
]


@app.post("/analyze")
async def analyze_image(image: UploadFile = File(...)) -> JSONResponse:
    """
    Analyze uploaded image for cat presence and activity.
    
    Returns:
        JSON response with:
        - cat_detected: bool
        - activity: str (one of the predefined activities)
        - confidence: float (0.0 to 1.0)
        - bounding_box: dict (if cat detected)
        - timestamp: str
    """
    try:
        # Read image data
        image_data = await image.read()
        
        # Convert to OpenCV format
        nparr = np.frombuffer(image_data, np.uint8)
        img = cv2.imdecode(nparr, cv2.IMREAD_COLOR)
        
        if img is None:
            raise HTTPException(status_code=400, detail="Invalid image format")
        
        # Perform analysis
        result = await perform_cat_analysis(img)
        
        return JSONResponse(content=result)
        
    except Exception as e:
        logger.error(f"Error analyzing image: {str(e)}")
        raise HTTPException(status_code=500, detail=f"Analysis failed: {str(e)}")


async def perform_cat_analysis(img: np.ndarray) -> Dict[str, Any]:
    """
    Perform cat detection and activity analysis on the image.
    
    This is where you would integrate with your actual AI/ML model.
    """
    
    # METHOD 1: Using a pre-trained YOLO model (uncomment if you have YOLO installed)
    """
    results = model(img)
    
    cat_detected = False
    confidence = 0.0
    bounding_box = None
    activity = "unknown"
    
    for result in results:
        boxes = result.boxes
        if boxes is not None:
            for box in boxes:
                # Check if detected object is a cat (class 15 in COCO dataset)
                if int(box.cls) == 15:  # Cat class in COCO
                    cat_detected = True
                    confidence = float(box.conf)
                    
                    # Get bounding box coordinates
                    x1, y1, x2, y2 = box.xyxy[0].tolist()
                    bounding_box = {
                        "x1": int(x1), "y1": int(y1),
                        "x2": int(x2), "y2": int(y2),
                        "width": int(x2 - x1),
                        "height": int(y2 - y1)
                    }
                    
                    # Analyze activity based on pose/position
                    activity = analyze_cat_activity(img, bounding_box)
                    break
    """
    
    # METHOD 2: Using cloud AI services (example with placeholder)
    """
    # For AWS Rekognition:
    import boto3
    
    rekognition = boto3.client('rekognition')
    _, buffer = cv2.imencode('.jpg', img)
    image_bytes = buffer.tobytes()
    
    response = rekognition.detect_labels(
        Image={'Bytes': image_bytes},
        MaxLabels=10,
        MinConfidence=70
    )
    
    cat_detected = any(label['Name'].lower() in ['cat', 'pet', 'animal'] 
                      for label in response['Labels'])
    """
    
    # METHOD 3: Basic placeholder implementation (current)
    height, width = img.shape[:2]
    
    # Convert to grayscale for basic analysis
    gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
    
    # Use basic computer vision techniques (placeholder)
    # In reality, you'd use proper ML models
    edges = cv2.Canny(gray, 50, 150)
    contours, _ = cv2.findContours(edges, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
    
    # Very basic heuristic (replace with actual ML model)
    large_contours = [c for c in contours if cv2.contourArea(c) > 1000]
    
    if large_contours:
        # Find the largest contour (assume it might be a cat)
        largest_contour = max(large_contours, key=cv2.contourArea)
        x, y, w, h = cv2.boundingRect(largest_contour)
        
        # Basic shape analysis
        aspect_ratio = w / h
        area_ratio = cv2.contourArea(largest_contour) / (w * h)
        
        # Heuristic for cat-like shapes
        cat_detected = (0.5 < aspect_ratio < 2.5 and area_ratio > 0.3)
        confidence = min(0.8, area_ratio) if cat_detected else 0.1
        
        bounding_box = {
            "x1": int(x), "y1": int(y),
            "x2": int(x + w), "y2": int(y + h),
            "width": int(w), "height": int(h)
        } if cat_detected else None
        
        # Analyze activity based on shape characteristics
        activity = analyze_cat_activity_basic(aspect_ratio, area_ratio) if cat_detected else "unknown"
        
    else:
        cat_detected = False
        confidence = 0.0
        bounding_box = None
        activity = "unknown"
    
    return {
        "cat_detected": cat_detected,
        "activity": activity,
        "confidence": round(confidence, 3),
        "bounding_box": bounding_box,
        "timestamp": datetime.now().isoformat(),
        "image_size": {"width": width, "height": height},
        "analysis_method": "basic_cv"
    }


def analyze_cat_activity_basic(aspect_ratio: float, area_ratio: float) -> str:
    """
    Basic activity analysis based on shape characteristics.
    In a real implementation, this would use pose estimation or activity recognition models.
    """
    
    # Very basic heuristics (replace with actual activity recognition)
    if aspect_ratio > 1.8:  # Wide shape
        return "lying_down"
    elif aspect_ratio < 0.8:  # Tall shape
        return "sitting"
    elif area_ratio < 0.4:  # Irregular shape
        return "playing"
    else:
        return "exploring"


@app.get("/health")
async def health_check():
    """Health check endpoint."""
    return {"status": "healthy", "timestamp": datetime.now().isoformat()}


@app.get("/")
async def root():
    """Root endpoint with service information."""
    return {
        "service": "Cat Activities Analysis Service",
        "version": "1.0.0",
        "endpoints": {
            "analyze": "POST /analyze - Upload image for analysis",
            "health": "GET /health - Health check"
        },
        "supported_activities": CAT_ACTIVITIES
    }


if __name__ == "__main__":
    import uvicorn
    
    print("Starting Cat Activities Analysis Service...")
    print("Upload images to http://localhost:8000/analyze")
    print("Visit http://localhost:8000 for service info")
    
    uvicorn.run(app, host="0.0.0.0", port=8000) 