cat-activities-analysis-api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
python-multipart>=0.0.6

# Optional AI/ML libraries (uncomment as needed)
# torch>=2.0.0
# torchvision>=0.15.0
# tensorflow>=2.13.0
# ultralytics>=8.0.0  # For YOLO models
# boto3>=1.29.0  # For AWS services
# google-cloud-vision>=3.4.0  # For Google Cloud Vision 